+++ packages/views/projects/components/index.ts
@@ -2,4 +2,5 @@ export { ProjectsPage } from "./projects-page";
 export { ProjectDetail } from "./project-detail";
 export { ProjectPicker } from "./project-picker";
 export { ProjectChip } from "./project-chip";
+export { ProjectMentionCard } from "./project-mention-card";
 export { LocalDirectoryHint } from "./local-directory-hint";
